Upgrade vendored Mathjax version to 3.2.0 Trello:https://trello.com/c/FnjQZn9Z/
We want to display Mathjax as SVG's in the mobile app. To do that, we
need this library with a slight modification to how it imports certain packages [1]. To that end we have forked the repo. As we are making a change to the library anyway, we are taking the opportunity to update the version of MathJax that is vendored.

In order to upgrade and re-vendor Mathjax-src, we followed Seb's previous work[2]

1. checked out the master branch from https://github.com/mathjaxMathJax-src/
2. npm install && npm run compile
3. cleaned the mathjax/es5/js directory (fromreact-native-mathjax-svg)
4. copied MathJax-src/js/ to react-native-mathjax-svg/mathjax/es5/js

[1] futurelearn/mobile@ea7418e
[2] futurelearn/mobile@d88dc37

Co-authored-by: roberta <roberta.mataityte@futurelearn.com>