Fix import of MathJax packages
Trello:https://trello.com/c/FnjQZn9Z/

We want to display Mathjax as SVG's in the mobile app. To do that, we
need this library with a slight modification to how it imports certain packages.

Previously we were not able to render characters/commands[1] from the
different extension libraries to MathJax like `ams`[2] because they were not being imported correctly. This commit fixes that.

[1] https://docs.mathjax.org/en/v2.7-latest/tex.html#supported-latex-commands
[2] http://docs.mathjax.org/en/latest/input/tex/extensions/index.html

Co-authored-by: roberta <roberta.mataityte@futurelearn.com>